Handle SDL errors/return types with MayError
These can be used to either throw exceptions on errors or check return codes C
style. May be less verbose than using exceptions to handle SDL errors that are
not truly exceptional situations. We can also use this code in Emscripten
without needing exceptions turned on.

I hope that when we ignore the error message, the inliner can see through
the function definition and delete the unnecessary checks.