Add error handling, rename macro files, small cleanups
Apparently i forgot to commit for a while.

Switch to constexpr if instead of preprocessor if when possible.

Provide a type for returning error codes that you can either check manually or
you can ignore and let them throw an exception. The goal is for this library to
be exception neutral but allow errors to be propagated out of here with
exceptions if desired.

Make checks for SDL's error codes more explicit.

Constrain the arguments of wrappers around SDL's functions so I (or others) can
write sensible overloads without modifying the existing code.